+++ drivers/clock_control/clock_stm32f2x.c
@@ -1,6 +1,6 @@
 /*
  *
- * Copyright (c) 2019 Linaro Limited.
+ * Copyright (c) 2018 qianfan Zhao
  *
  * SPDX-License-Identifier: Apache-2.0
  */
@@ -11,17 +11,17 @@
 #include <clock_control.h>
 #include <misc/util.h>
 #include <clock_control/stm32_clock_control.h>
-#include "stm32_ll_clock.h"
+#include "clock_stm32_ll_common.h"
+
 
 #ifdef CONFIG_CLOCK_STM32_SYSCLK_SRC_PLL
 
 /* Macros to fill up division factors values */
-#define _pllm(v) LL_RCC_PLLM_DIV_ ## v
-#define pllm(v) _pllm(v)
+#define z_pllm(v) LL_RCC_PLLM_DIV_ ## v
+#define pllm(v) z_pllm(v)
 
-#define _pllr(v) LL_RCC_PLLR_DIV_ ## v
-#define pllr(v) _pllr(v)
-
+#define z_pllp(v) LL_RCC_PLLP_DIV_ ## v
+#define pllp(v) z_pllp(v)
 
 /**
  * @brief fill in pll configuration structure
@@ -30,7 +30,7 @@ void config_pll_init(LL_UTILS_PLLInitTypeDef *pllinit)
 {
 	pllinit->PLLM = pllm(CONFIG_CLOCK_STM32_PLL_M_DIVISOR);
 	pllinit->PLLN = CONFIG_CLOCK_STM32_PLL_N_MULTIPLIER;
-	pllinit->PLLR = pllr(CONFIG_CLOCK_STM32_PLL_R_DIVISOR);
+	pllinit->PLLP = pllp(CONFIG_CLOCK_STM32_PLL_P_DIVISOR);
 }
 #endif /* CONFIG_CLOCK_STM32_SYSCLK_SRC_PLL */
 
@@ -39,5 +39,14 @@ void config_pll_init(LL_UTILS_PLLInitTypeDef *pllinit)
  */
 void config_enable_default_clocks(void)
 {
-	/* Nothing for now */
+	/* Power Interface clock enabled by default */
+	LL_APB1_GRP1_EnableClock(LL_APB1_GRP1_PERIPH_PWR);
+}
+
+/**
+ * @brief Function kept for driver genericity
+ */
+void LL_RCC_MSI_Disable(void)
+{
+	/* Do nothing */
 }
