+++ drivers/clock_control/clock_stm32wbx.c
@@ -1,25 +1,27 @@
 /*
- * Copyright (c) 2018 Yurii Hamann
+ *
+ * Copyright (c) 2019 Linaro Limited.
  *
  * SPDX-License-Identifier: Apache-2.0
  */
+
 
 #include <soc.h>
 #include <soc_registers.h>
 #include <clock_control.h>
 #include <misc/util.h>
 #include <clock_control/stm32_clock_control.h>
-#include "stm32_ll_clock.h"
-
+#include "clock_stm32_ll_common.h"
 
 #ifdef CONFIG_CLOCK_STM32_SYSCLK_SRC_PLL
 
 /* Macros to fill up division factors values */
-#define z_pllm(v) LL_RCC_PLLM_DIV_ ## v
-#define pllm(v) z_pllm(v)
+#define _pllm(v) LL_RCC_PLLM_DIV_ ## v
+#define pllm(v) _pllm(v)
 
-#define z_pllp(v) LL_RCC_PLLP_DIV_ ## v
-#define pllp(v) z_pllp(v)
+#define _pllr(v) LL_RCC_PLLR_DIV_ ## v
+#define pllr(v) _pllr(v)
+
 
 /**
  * @brief fill in pll configuration structure
@@ -28,7 +30,7 @@ void config_pll_init(LL_UTILS_PLLInitTypeDef *pllinit)
 {
 	pllinit->PLLM = pllm(CONFIG_CLOCK_STM32_PLL_M_DIVISOR);
 	pllinit->PLLN = CONFIG_CLOCK_STM32_PLL_N_MULTIPLIER;
-	pllinit->PLLP = pllp(CONFIG_CLOCK_STM32_PLL_P_DIVISOR);
+	pllinit->PLLR = pllr(CONFIG_CLOCK_STM32_PLL_R_DIVISOR);
 }
 #endif /* CONFIG_CLOCK_STM32_SYSCLK_SRC_PLL */
 
@@ -37,14 +39,5 @@ void config_pll_init(LL_UTILS_PLLInitTypeDef *pllinit)
  */
 void config_enable_default_clocks(void)
 {
-	/* Power Interface clock enabled by default */
-	LL_APB1_GRP1_EnableClock(LL_APB1_GRP1_PERIPH_PWR);
-}
-
-/**
- * @brief Function kept for driver genericity
- */
-void LL_RCC_MSI_Disable(void)
-{
-	/* Do nothing */
+	/* Nothing for now */
 }
